add improvement plan tab
